pyproject.toml
[tool.poetry]
name = "Kilobyte"
version = "0.1.0"
description = "projeto final senai API"
authors = ["Henrique Messias"]
license = "The Unlicense"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.0"
mysql-connector-python = "^8.2.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
